Fix bug where comparing 32-bit values in global arrays or structs against 0 may give wrong results with large memory model.
The issue was that 16-bit absolute addressing (in the data bank) was being used to access the data to compare, but with the large memory model the global arrays or structs are not necessarily in the same bank, so absolute long addressing should be used.

This corresponds to byteworksinc/ORCA-C@709f9b3f25c7beb6ded5fbaca560c0fa0a49f468.

The following program often demonstrates the problem (depending on memory layout and contents):

{$memorymodel 1}
{$optimize 1}
program largemmbug(output);

var
        ch1: packed array[1..32000] of char;
        L1: array[0..1] of longint;

begin
L1[0] := 0;
if L1[0] <> 0 then
        writeln('bug detected'); {depends on memory layout and contents}
end.