Fix bug causing functions with 254 bytes of locals and parameters to crash on return.
This bug occurred because the generated code tried to store part of the return address to a direct page offset of 256, but instead an offset of 0 was used, resulting in an invalid return address and (typically) a crash. It could occur if the function took one or more parameters, and the total size of parameters and local variables (including compiler-generated ones) was 254 bytes.

This corresponds to byteworksinc/ORCA-C@fb612e14d15788d57529be02410570768c2f1a51.

The following program demonstrates the problem:

program biglocals1;

var
        x: integer;

function testfn(a: integer; b: longint): integer;
var
        s: packed array[0..245] of char;
begin
testfn := a;
end;

begin
x := testfn(1,2);
end.