Fix bug causing functions with 254 bytes of locals to crash on return.
This was a bug with the code for moving the return address. It would generate a "LDA 0" instruction when it was trying to load the value at DP+256.

This corresponds to byteworksinc/ORCA-C@29de86703927170785c59fa2c55ecc5bb624ceb2.

The following program demonstrates the crash:

program local254;

procedure testproc(x: integer);
var
        a: array[0..253] of byte;
begin
end;

begin
testproc(1);
end.