Fix code generation for certain cases where addresses are stored to local variables that don't fit in the direct page.
There was a bug when storing addresses generated by expressions like @a[i], where a is a global array and i is a variable. In certain cases where the destination location was a local variable that didn't fit in the direct page, the result of the address calculation would be stored to the wrong location on the stack. This failed to give the correct result, and could also sometimes cause crashes or other problems due to stack corruption.

This corresponds to byteworksinc/ORCA-C@fd0ff033adc55bd2afb8752d24e22a9630d14b7d, but the code change is simplified from what was used there. (The ORCA-C patch includes code to handle pc_cop intermediate codes, but that should never actually be needed.)

The following program illustrates the issues:

{$optimize 1}
program storeaddr(output);

var
        a: array [0..4] of longint;
        idx: integer;

function testfn: longint;
var
        b: array [0..70] of ^longint;
begin
        b[70] := @a[idx];
        testfn := b[70]^;
end;

begin
idx := 4;
a[4] := 123;
writeln(testfn); {should print 123}
end.