Fix evaluation of conditionals involving shifts.
The code for 16-bit shifts was assumed to set the z flag based on the result value, but did not actually do so. This could result in mis-evaluation of conditionals.

This corresponds to byteworksinc/ORCA-C@7898c619c8af29ad4f319e68c38f468a7b3731f7 with regard to shifts. (The ORCA/C patch also fixed cases involving bitfields or the ?: operator, which do not apply to Pascal.)

Here is an example showing the problem:

program shiftcond(output);

var
        i,j: integer;

begin
i := 1;
j := 0;
if (i >> j) <> 0 then
        writeln('OK');
end.