Fix bug with 4-byte comparisons against globals in large memory model.
Long addressing was not being used to access the values, which could lead to mis-evaluation of comparisons against values in global records or arrays, depending on the memory layout.

This corresponds to byteworksinc/ORCA-C@25500815179508d2e7eb01846468db94274ff3ca.

Here is an example that can be affected by this (depending on memory layout):

{$memorymodel 1}
program globalcmp(output);

type
        recPtr = ^theRec;
        theRec = record
                p: recPtr;
        end;

var
        s: theRec;

begin
        s.p := @s;
        writeln(s.p <> @s); {should be false}
end.